I have written certain benchmark codes and found a shocking surprise
Csq code:
```
i := 0
while i < 5000:
 i = i+1
```
Python code:
```
i = 0
while i < 5000:
 i = i+1
```

THe results were:
Csq took 159ms
Python took 71ms

I then searched for the code decreasing speed and try to optimize it and now the results are:
Csq took 13ms
Python took 61ms